# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API project for managing contacts",
    "version": "1.0.0"
  },
  "host": "cse341-bh1k.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api-docs/swagger.json": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "Get all contacts or a specific contact by ID using query parameter",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "description": "Optional contact ID to retrieve a specific contact"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved contact(s)"
          },
          "204": {
            "description": "No contacts found"
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Contact information to create",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@email.com"
                },
                "favoriteColor": {
                  "type": "string",
                  "example": "Blue"
                },
                "birthday": {
                  "type": "string",
                  "example": "1990-05-15"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact successfully created"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid request body or missing required fields"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Database service unavailable"
          }
        }
      }
    },
    "/contacts/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    }
  }
}